requirements.txt
fastapi
pyserial
jinja2
pytest
